Fixed random generator, moved user agents to seperate module to make a private field
